fail job immediately if it cannot be fit due to walltime > maxtime
git-svn-id: https://svn.code.sf.net/p/cogkit/svn/branches/4.1.9/src/cog@3220 5b74d2a0-fa0e-0410-85ed-ffba77ec0bde

+++ modules/provider-coaster/src/org/globus/cog/abstraction/coaster/service/job/manager/TimeInterval.java
@@ -72,7 +72,7 @@ public final class TimeInterval implements Comparable<TimeInterval> {
         return ms;
     }
     
-    public static TimeInterval fromSeconds(int seconds) {
+    public static TimeInterval fromSeconds(long seconds) {
         return new TimeInterval(seconds * 1000);
     }
